move the url normalization function to utils
